% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recountWorkflow-package.R
\docType{package}
\name{recountWorkflow-package}
\alias{recountWorkflow}
\alias{recountWorkflow-package}
\title{recountWorkflow: recount workflow: accessing over 70,000 human RNA-seq samples with
    Bioconductor}
\description{
The recount2 resource is composed of over 70,000 uniformly
    processed human RNA-seq samples spanning TCGA and SRA, including GTEx. The
    processed data can be accessed via the recount2 website and the recount 
    Bioconductor package. This workflow explains in detail how to use the 
    recount package and how to integrate it with other Bioconductor packages for
    several analyses that can be carried out with the recount2 resource. In
    particular, we describe how the coverage count matrices were computed in
    recount2 as well as different ways of obtaining public metadata, which can
    facilitate downstream analyses. Step-by-step directions show how to do a
    gene level differential expression analysis, visualize base-level genome
    coverage data, and perform an analyses at multiple feature levels. This
    workflow thus provides further information to understand the data in
    recount2 and a compendium of R code to use the data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LieberInstitute/recountWorkflow}
  \item Report bugs at \url{https://support.bioconductor.org/t/recountWorkflow/}
}

}
\author{
\strong{Maintainer}: Leonardo Collado-Torres \email{lcolladotor@gmail.com}

Other contributors:
\itemize{
  \item Abhinav Nellore \email{anellore@gmail.com} [contributor]
  \item Andrew E. Jaffe \email{andrew.jaffe@libd.org} [contributor]
}

}
\keyword{internal}
